requirements.txt
autopep8
celery
coverage
django-debug-toolbar
django-extensions
djangorestframework-simplejwt
pep8
pip-chill
psycopg2
python-decouple
redis
zeep
